[ML] Filtering category runtime fields in advanced wizard (#132916) (#132987)
* [ML] Filtering category runtime fields in advanced wizard

* reverting accidentally commited code

* sorting fields in dropdowns

* removing accidentally committed code

(cherry picked from commit ad36a63469788dd957a4e9b70746db624c6c6e56)
