+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/by_field/by_field.tsx
@@ -10,11 +10,9 @@ import { i18n } from '@kbn/i18n';
 
 import { SplitFieldSelect } from '../split_field_select';
 import { JobCreatorContext } from '../../../job_creator_context';
-import { Field } from '../../../../../../../../../common/types/fields';
-import {
-  newJobCapsService,
-  filterCategoryFields,
-} from '../../../../../../../services/new_job_capabilities/new_job_capabilities_service';
+import type { Field } from '../../../../../../../../../common/types/fields';
+import { filterCategoryFields } from '../../../../../../../../../common/util/fields_utils';
+import { newJobCapsService } from '../../../../../../../services/new_job_capabilities/new_job_capabilities_service';
 import { PopulationJobCreator } from '../../../../../common/job_creator';
 
 interface Props {
@@ -27,7 +25,10 @@ export const ByFieldSelector: FC<Props> = ({ detectorIndex }) => {
 
   const runtimeCategoryFields = useMemo(() => filterCategoryFields(jobCreator.runtimeFields), []);
   const allCategoryFields = useMemo(
-    () => [...newJobCapsService.categoryFields, ...runtimeCategoryFields],
+    () =>
+      [...newJobCapsService.categoryFields, ...runtimeCategoryFields].sort((a, b) =>
+        a.name.localeCompare(b.name)
+      ),
     []
   );
