+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/rare_field/rare_field.tsx
@@ -9,12 +9,10 @@ import React, { FC, useContext, useEffect, useState, useMemo } from 'react';
 
 import { RareFieldSelect } from './rare_field_select';
 import { JobCreatorContext } from '../../../job_creator_context';
-import {
-  newJobCapsService,
-  filterCategoryFields,
-} from '../../../../../../../services/new_job_capabilities/new_job_capabilities_service';
+import { filterCategoryFields } from '../../../../../../../../../common/util/fields_utils';
+import { newJobCapsService } from '../../../../../../../services/new_job_capabilities/new_job_capabilities_service';
 import { Description } from './description';
-import { Field } from '../../../../../../../../../common/types/fields';
+import type { Field } from '../../../../../../../../../common/types/fields';
 import { RareJobCreator } from '../../../../../common/job_creator';
 
 export const RareFieldSelector: FC = () => {
@@ -23,7 +21,10 @@ export const RareFieldSelector: FC = () => {
 
   const runtimeCategoryFields = useMemo(() => filterCategoryFields(jobCreator.runtimeFields), []);
   const allCategoryFields = useMemo(
-    () => [...newJobCapsService.categoryFields, ...runtimeCategoryFields],
+    () =>
+      [...newJobCapsService.categoryFields, ...runtimeCategoryFields].sort((a, b) =>
+        a.name.localeCompare(b.name)
+      ),
     []
   );
   const categoryFields = useFilteredCategoryFields(
