+++ x-pack/plugins/ml/public/application/services/new_job_capabilities/new_job_capabilities_service.ts
@@ -5,23 +5,12 @@
  * 2.0.
  */
 import { ES_FIELD_TYPES } from '@kbn/data-plugin/public';
-import { DataView } from '@kbn/data-views-plugin/public';
-import {
-  Field,
-  Aggregation,
-  AggId,
-  FieldId,
-  EVENT_RATE_FIELD_ID,
-} from '../../../../common/types/fields';
+import type { DataView } from '@kbn/data-views-plugin/public';
+import type { Field, Aggregation, AggId, FieldId } from '../../../../common/types/fields';
+import { EVENT_RATE_FIELD_ID } from '../../../../common/types/fields';
+import { filterCategoryFields } from '../../../../common/util/fields_utils';
 import { ml } from '../ml_api_service';
 import { processTextAndKeywordFields, NewJobCapabilitiesServiceBase } from './new_job_capabilities';
-
-const categoryFieldTypes = [
-  ES_FIELD_TYPES.TEXT,
-  ES_FIELD_TYPES.KEYWORD,
-  ES_FIELD_TYPES.IP,
-  ES_FIELD_TYPES.VERSION,
-];
 
 class NewJobCapsService extends NewJobCapabilitiesServiceBase {
   private _catFields: Field[] = [];
@@ -170,10 +159,6 @@ function addEventRateField(aggs: Aggregation[], fields: Field[]) {
     }
   });
   fields.splice(0, 0, eventRateField);
-}
-
-export function filterCategoryFields(fields: Field[]) {
-  return fields.filter((f) => categoryFieldTypes.includes(f.type));
 }
 
 export const newJobCapsService = new NewJobCapsService();
